% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{beta_est.default}
\alias{beta_est.default}
\title{An S3 method for getting Beta for a model}
\usage{
\method{beta_est}{default}(squire_model, model_params, R0)
}
\arguments{
\item{squire_model}{A model object}

\item{model_params}{Parameters for the model}

\item{R0}{R0/Rt value to translate into beta}
}
\description{
An S3 method for getting Beta for a model
}
