% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bhu_fit}
\alias{bhu_fit}
\title{Nimue fit to excess mortality for Bhutan}
\format{
Nimue Simulation Object:
\describe{
\item{parameters}{list of fixed parameters}
\item{model}{the odin model}
\item{odin_parameters}{list of fixed parameters in the format needed for the odin model}
\item{replicate_parameters}{data-frame of a sample of the parameters estimated by the MCMC}
\item{pmcmc_results}{full matrix of parameter estimates generated by the pmcmc}
\item{interventions}{various other parameters used in the model, i.e. to generate Rt over-time from the replicates}
}
}
\usage{
bhu_fit
}
\description{
Included for use in examples. Fitted to data up to 2021-11-16.
}
\keyword{datasets}
