% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{block_interpolate}
\alias{block_interpolate}
\title{Convert a vector of parameter values and a vector of the times it changes into a time-series of the parameter values.}
\usage{
block_interpolate(t, x, tt_x)
}
\arguments{
\item{t}{Vector of times we want parameter values for.}

\item{x}{Vector (list or matrix or array) of parameter values that correspond to each tt_x.}

\item{tt_x}{Vector of times at which the parameter x changes.}
}
\value{
A vector of values for parameter x that correspond to the times specified by t
}
\description{
This essentially replicates the interpolate(_, "constant") function of odin, but returns a time series of values at the requested times t.
The implementation for 1D vector x's is not finished and should be checked before usage, alternatively x can be replaced with indexes from x's time dimension.
}
