% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cdp_plot}
\alias{cdp_plot}
\title{Plot the cumulative deaths from the model}
\usage{
cdp_plot(res, extra_df = NULL)
}
\arguments{
\item{res}{model output}

\item{extra_df}{data to plot if need}
}
\description{
Plot the cumulative deaths from the model
}
