% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compare_adjustment_plot}
\alias{compare_adjustment_plot}
\title{A diagnostic plot for looking at the infections-based fitting}
\usage{
compare_adjustment_plot(out)
}
\arguments{
\item{out}{model object}
}
\description{
A diagnostic plot for looking at the infections-based fitting
}
