% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{country_immunity_plot}
\alias{country_immunity_plot}
\title{Plot the Rt trends for a country accounting for immunity}
\usage{
country_immunity_plot(
  df,
  min_date,
  date_0,
  vjust = -1.2,
  R0 = FALSE,
  Rt = FALSE,
  Reff = TRUE
)
}
\arguments{
\item{df}{model output}

\item{min_date}{Date to start the plot at}

\item{date_0}{Current date}

\item{vjust}{unused}

\item{R0}{plot R0}

\item{Rt}{plot Rt}

\item{Reff}{plot Reff}
}
\description{
Plot the Rt trends for a country accounting for immunity
}
