% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dp_plot}
\alias{dp_plot}
\title{Plot daily model deaths.}
\usage{
dp_plot(res)
}
\arguments{
\item{res}{A squire/nimue model object with generated fits (i.e. output of
generate_draws)}
}
\value{
A ggplot2 object
}
\description{
Returns a ggplot2 object with median (95\% CI) modelled deaths and the data it
has been fitted too.
}
