% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections_extend_inputs.R
\name{extend_vaccine_inputs}
\alias{extend_vaccine_inputs}
\title{Updates vaccine inputs into the future.
I.e. prevents coverage of second doses going above 100\% of first doses.}
\usage{
extend_vaccine_inputs(vacc_inputs, time_period, out, end_date)
}
\arguments{
\item{vacc_inputs}{inputs used in the model}

\item{time_period}{how far to model}

\item{out}{model object}

\item{end_date}{date that our fit ends on}
}
\description{
Updates vaccine inputs into the future.
I.e. prevents coverage of second doses going above 100\% of first doses.
}
