% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_draws.default}
\alias{generate_draws.default}
\title{Generate Draws from pmcmc run}
\usage{
\method{generate_draws}{default}(out, pars.list, draws, ...)
}
\arguments{
\item{out}{Output of \code{squire::pmcmc}}

\item{pars.list}{Output of generate_parameters(), default = NULL calls
generate_parameters.}

\item{draws}{Number of draws from mcmc chain. Default = 10, if NULL then uses
replicate_parameters as the parameters}

\item{...}{method specific arguments, unused}
}
\description{
Uses furrr, so can be called in parallel.
}
