% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_draws}
\alias{generate_draws}
\title{Generate Draws from pmcmc run}
\usage{
generate_draws(
  out,
  pars.list = NULL,
  draws = 10,
  parallel = TRUE,
  noInfectionProtect = FALSE
)
}
\arguments{
\item{out}{Output of \code{squire::pmcmc}}

\item{pars.list}{Output of generate_parameters(), default = NULL calls
generate_parameters.}

\item{draws}{Number of draws from mcmc chain. Default = 10, if NULL then uses
replicate_parameters as the parameters}

\item{parallel}{Are simulations done in parallel? Default = TRUE}

\item{noInfectionProtect}{Should the vaccine prevent infections. Default = FALSE}
}
\description{
Generate Draws from pmcmc run
}
