% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_excess.R
\name{generate_draws_counterfactual}
\alias{generate_draws_counterfactual}
\title{Generate Draws from pmcmc run with a counter factual based on vaccines}
\usage{
generate_draws_counterfactual(
  out,
  pars.list,
  assignedVaccine,
  vaccineStart = NULL,
  draws = 10,
  parallel = TRUE,
  excess = excess
)
}
\arguments{
\item{out}{Output of \code{squire::pmcmc}}

\item{pars.list}{Output of generate_parameters().}

\item{assignedVaccine}{The number of doses given to the country at this date.}

\item{vaccineStart}{The date the vaccination is to start. Should be a character
in the format "YYYY-MM-DD". Default = NULL}

\item{draws}{Number of draws from mcmc chain. Default = 10}

\item{parallel}{Are simulations done in parallel? Default = TRUE}

\item{excess}{Use the excess loglikelihood? Default = FALSE}
}
\description{
Generate Draws from pmcmc run with a counter factual based on vaccines
}
