% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_fitting.R
\name{generate_draws_pmcmc_nimue_case_fitted}
\alias{generate_draws_pmcmc_nimue_case_fitted}
\title{Scale a Nimue MCMC output by the number of reported cases}
\usage{
generate_draws_pmcmc_nimue_case_fitted(out, n_particles = 10, grad_dur = 21)
}
\arguments{
\item{out}{A nimue model object.}

\item{n_particles}{How many replicates to draw from the adjust object.}

\item{grad_dur}{How many days into the past to adjust.}
}
\value{
A nimue model object, with the adjusted replicate parameters.
}
\description{
Performs a simple grid search, adding or subtracting from the final fitted
effect (in all replicates of the model object). Chooses the value that
minimises the error from the total number of infections in the final
\code{grad_dur} days of the epidemic, adjusted by a reporting fraction
calculated over the previous few days (14 by default, determined by
\code{Rt_rw_duration}). A legacy function and is no longer used.
}
