% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Generate parameter draws from a pmcmc run
Code take from squire:::sample_pmcmc()}
\usage{
generate_parameters(out, draws = 10)
}
\arguments{
\item{out}{Output of \code{squire::pmcmc}}

\item{draws}{Number of draws from mcmc chain. Default = 10}
}
\description{
Generate parameter draws from a pmcmc run
Code take from squire:::sample_pmcmc()
}
