% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3cs.R
\name{get_covax_iso3c}
\alias{get_covax_iso3c}
\title{Get iso3 codes of recipient countries in COVAX}
\usage{
get_covax_iso3c()
}
\value{
Vector of iso3 codes.
}
\description{
\url{https://www.who.int/initiatives/act-accelerator/covax}
}
