% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_data.particle_fit}
\alias{get_data.particle_fit}
\title{S3 Method to get all the data of a model}
\usage{
\method{get_data}{particle_fit}(model_out)
}
\description{
S3 Method to get all the data of a model
}
