% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_data_end_date.excess_nimue_simulation}
\alias{get_data_end_date.excess_nimue_simulation}
\title{S3 Method to get final date in the data from a model}
\usage{
\method{get_data_end_date}{excess_nimue_simulation}(model_out)
}
\description{
S3 Method to get final date in the data from a model
}
