% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_data_end_date_inner.particle_fit}
\alias{get_data_end_date_inner.particle_fit}
\title{S3 Method to get start date of the final time-period in the data from a model}
\usage{
\method{get_data_end_date_inner}{particle_fit}(model_out)
}
\description{
S3 Method to get start date of the final time-period in the data from a model
}
