% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_data_end_date_inner.rt_optimised}
\alias{get_data_end_date_inner.rt_optimised}
\title{S3 Method to get start date of the final time-period in the data from a model}
\usage{
\method{get_data_end_date_inner}{rt_optimised}(model_out)
}
\arguments{
\item{model_out}{A nimue/squire mcmc or particle fit output}
}
\description{
S3 Method to get start date of the final time-period in the data from a model
}
