% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_data_start_date.nimue_simulation}
\alias{get_data_start_date.nimue_simulation}
\title{S3 Method to get the start date of the data in a model}
\usage{
\method{get_data_start_date}{nimue_simulation}(model_out)
}
\arguments{
\item{model_out}{A nimue/squire mcmc or particle fit output}
}
\description{
S3 Method to get the start date of the data in a model
}
