% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_dates.excess_nimue_simulation}
\alias{get_dates.excess_nimue_simulation}
\title{S3 Method to get all the start dates of time-periods in the data of a model}
\usage{
\method{get_dates}{excess_nimue_simulation}(model_out)
}
\description{
S3 Method to get all the start dates of time-periods in the data of a model
}
