% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_dates_greater.rt_optimised}
\alias{get_dates_greater.rt_optimised}
\title{S3 Method to get all the end dates of time-periods in the data of a model}
\usage{
\method{get_dates_greater}{rt_optimised}(model_out)
}
\arguments{
\item{model_out}{A nimue/squire mcmc or particle fit output}
}
\description{
S3 Method to get all the end dates of time-periods in the data of a model
}
