% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections_estimate_scenarios.R
\name{get_future_Rt}
\alias{get_future_Rt}
\title{Calculates the Rt used in the short term scenarios}
\usage{
get_future_Rt(model_out, forcast_days)
}
\arguments{
\item{model_out}{a model object}

\item{forcast_days}{how far to calculate for}
}
\description{
Calculates the Rt used in the short term scenarios
}
