% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3cs.R
\name{get_gavi_iso3c}
\alias{get_gavi_iso3c}
\title{Get iso3 codes of recipient countries in GAVI}
\source{
\url{https://www.gavi.org/}
}
\usage{
get_gavi_iso3c()
}
\value{
Vector of iso3 codes.
}
\description{
Get iso3 codes of recipient countries in GAVI
}
