% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3cs.R
\name{get_global_fund_iso3c}
\alias{get_global_fund_iso3c}
\title{Get iso3 codes of recipient countries in TheGlobalFund}
\source{
\url{https://www.theglobalfund.org/en/}
}
\usage{
get_global_fund_iso3c()
}
\value{
Vector of iso3 codes.
}
\description{
Get iso3 codes of recipient countries in TheGlobalFund
}
