% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_immunity_ratios.rt_optimised}
\alias{get_immunity_ratios.rt_optimised}
\title{An S3 method to calculate the proportion of the population immune}
\usage{
\method{get_immunity_ratios}{rt_optimised}(out, max_date = NULL, vaccine = FALSE)
}
\arguments{
\item{out}{squire/nimue model object}

\item{max_date}{date to generate immunity ratio upto, default = NULL gets the max_date from out's data}

\item{vaccine}{should the function account for vaccinations, default = FALSE}
}
\description{
An S3 method to calculate the proportion of the population immune
}
