% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3cs.R
\name{get_income_group}
\alias{get_income_group}
\title{Get the World Bank Income Categories for a given set of iso3 codes}
\source{
\url{https://datahelpdesk.worldbank.org/knowledgebase/articles/906519-world-bank-country-and-lending-groups}
}
\usage{
get_income_group(iso3cs)
}
\arguments{
\item{iso3cs}{Vector of iso 3 codes.}
}
\value{
Ordered factor of classifications codes.
}
\description{
Note that some countries that are currently unclassified have been given either
their previous classification (i.e Venezuela) or the classification of the country
that they are a territory of (i.e. Hong Kong).
}
