% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_likelihood.R
\name{get_model_likelihood.vacc_durR_nimue_simulation}
\alias{get_model_likelihood.vacc_durR_nimue_simulation}
\title{Method to calculate log-likelihood for each model type for a nimue model using excess deaths with varying vaccine duration}
\usage{
\method{get_model_likelihood}{vacc_durR_nimue_simulation}(model_out)
}
\arguments{
\item{model_out}{Model object}
}
\description{
Method to calculate log-likelihood for each model type for a nimue model using excess deaths with varying vaccine duration
}
