% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_parameters.rt_optimised}
\alias{get_parameters.rt_optimised}
\title{S3 Method to get total susceptible population}
\usage{
\method{get_parameters}{rt_optimised}(model_out)
}
\arguments{
\item{model_out}{A nimue/squire mcmc or particle fit output}
}
\description{
S3 Method to get total susceptible population
}
