% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{get_Rt}
\alias{get_Rt}
\alias{get_Rt.nimue_simulation}
\alias{get_Rt.excess_nimue_simulation}
\alias{get_Rt.vacc_durR_nimue_simulation}
\alias{get_Rt.rt_optimised}
\title{Methods to calculate Rt over time from each model type}
\usage{
get_Rt(model_out)

\method{get_Rt}{nimue_simulation}(model_out)

\method{get_Rt}{excess_nimue_simulation}(model_out)

\method{get_Rt}{vacc_durR_nimue_simulation}(model_out)

\method{get_Rt}{rt_optimised}(model_out)
}
\arguments{
\item{model_out}{Model object}
}
\description{
Methods to calculate Rt over time from each model type
}
