% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3cs.R
\name{get_WHO_region}
\alias{get_WHO_region}
\title{Get the WHO Regions for a given set of iso3 codes}
\source{
\url{https://www.who.int/countries}
}
\usage{
get_WHO_region(iso3cs)
}
\arguments{
\item{iso3cs}{Vector of iso 3 codes.}
}
\value{
Vector of WHO regions.
}
\description{
Note that some countries that are currently unclassified have been given the
classification of the country that they are a territory of (i.e. Hong Kong).
}
