% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booster_model.R
\name{nimue_booster_model}
\alias{nimue_booster_model}
\title{nimue model creation.}
\usage{
nimue_booster_model(use_dde = TRUE)
}
\arguments{
\item{use_dde}{Logical for using dde to solve. Default = TRUE
We will use this structure to ensure that model fitting is flexible in the
future as more models are added}
}
\description{
Create a nimue model class for fitting with squire tools using the booster
dose framework.
}
\details{
Wraps the squire pmcmc fitting infrastructure.
}
