% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimue_utils.R
\name{nimue_format}
\alias{nimue_format}
\title{Nimue format that returns just compartments/summaries requested}
\usage{
nimue_format(
  out,
  var_select = NULL,
  reduce_age = TRUE,
  combine_compartments = TRUE,
  date_0 = NULL,
  mean_over_80_age = 82.5,
  case_to_infection_ratio = 0.195
)
}
\arguments{
\item{out}{nimue_simulation or squire_simulation object}

\item{var_select}{Vector of compartment names, e.g. \code{c("S", "R")}. In
addition a number of summary compartment can be requested. These include:
\itemize{
\item{"deaths"}{ Daily Deaths }
\item{"infections"}{ Daily Infections }
\item{"hospital_occupancy"}{ See description }
\item{"ICU_occupancy"}{ See description }
\item{"hospital_demand"}{ See description }
\item{"ICU_demand"}{ See description }
\item{"hospital_incidence"}{ See description }
\item{"ICU_incidence"}{ See description }
\item{"hospitalisations"}{ See description }
\item{"vaccines", "unvaccinated", "vaccinated", "priorvaccinated"}{ Vaccine outputs }
\item{"long_covid"}{ Long COVID estimates}
}}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{combine_compartments}{Collapse compartments of same type together
(e.g. E1 and E2 -> E)}

\item{date_0}{Date of time 0, if specified a date column will be added}

\item{mean_over_80_age}{Mean age of over 80 population for calculating long
covid incidence. Default = 82.5}

\item{case_to_infection_ratio}{Ratio of infections to cases for long covid
calculation. Default = 0.195 from UK ONS analyses}
}
\value{
Formatted long data.frame
}
\description{
If a squire object is provided out it should be feed through to
\code{squire::format_output}. However, please note that not all variables can
be selected for squire.
}
\details{
To better define the summaries related to the healthcare pathway:
\describe{
\item{demand}{The total number of requiring hospitalisation and ICU care
at the given time, regardless of if they've received it}
\item{occupancy}{The total number of those who received ICU or hospital care
at the given time}
\item{incidence}{The total number of new people who require hospital or ICU
care}
\item{hospitalisations}{Sum of new people recieving hospital or ICU care}
}
Note: this is confusing and seemingly inconsistent i.e. hospital_incidence
does not capture those recovering from ICU and hospitalisations is stated
to equal the sum of the incidences elsewhere.
}
\examples{
#generate some outputs for the afghanistan fit and pass through to format
#getting deaths and long covid estimates
\dontrun{
generate_draws(afg_fit) \%>\%
  nimue_format(var_select = c("deaths", "long_covid"))
}
}
