% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_fit.R
\name{particle_fit}
\alias{particle_fit}
\title{Fit nimue to a given set of death data via the particle exploration approach.}
\usage{
particle_fit(
  data,
  distribution,
  squire_model,
  parameters,
  start_date,
  rt_spacing = 14,
  initial_r = 3,
  proposal_width = 0.5,
  n_particles = 7,
  k = 14
)
}
\arguments{
\item{data}{A data frame of deaths occuring over a given time frame. Given in the
format: deaths(integer), date_start(date), date_end(date). Must have at least
one death period in each set of Rt trend changes (i.e. a 14 day period by
default) and no period can overlap these changes.}

\item{distribution}{A list of samples(a list) with names specifying parameters.}

\item{squire_model}{A model object of the desired type to use, i.e. squire, nimue.}

\item{parameters}{A list of parameters to keep the same across all samples.}

\item{start_date}{Date when the epidemic begins, parameters and distribution
should be formatted relative to this date, where necessary.}

\item{rt_spacing}{Number of days between each Rt trend, default = 14 days.}

\item{initial_r}{The value of R0 to centre on initially, default = 3.}

\item{proposal_width}{What proportion of the previous Rt trend to explore, default = 50\%.}

\item{n_particles}{How many particles to explore uniformly across that range, default = 7.}

\item{k}{Control the dispersion on the negative binomial likelihood, default = 14.}
}
\value{
An object of type particle_fit, (model type).
}
\description{
For a given number of samples from given parameter uncertainty or distribution
function, this approach iteratively fits the Rt trend to the provided death
data and returns a nimue_simulation object for future usage in scenario
modelling.
}
