% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{plot.particle_fit}
\alias{plot.particle_fit}
\title{Essentially just does squire::plot_pmcmc_sample. Not intended for use otherwise.}
\usage{
\method{plot}{particle_fit}(
  x,
  q = c(0.025, 0.975),
  replicates = TRUE,
  summarise = FALSE,
  ci = TRUE,
  ...
)
}
\arguments{
\item{x}{particle_fit output object}

\item{q}{Quantiles for upper and lower of interval ribbon}

\item{replicates}{Plot replicates}

\item{summarise}{Logical, add summary line}

\item{ci}{logical add confidence interval ribbon}

\item{...}{placeholder for compatibility does nothing.}
}
\description{
Essentially just does squire::plot_pmcmc_sample. Not intended for use otherwise.
}
