% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comet.R
\name{prepare_input_json_df}
\alias{prepare_input_json_df}
\title{Convert a model output into a data frame for Comet/GAVI Analysis}
\usage{
prepare_input_json_df(out, ox_interventions = NULL)
}
\arguments{
\item{out}{Nimue or squire model output, must have MCMC results in
\code{pmcmc_results} slot.}

\item{ox_interventions}{Data frame of government interventions, used to add
graphical elements, if \code{NULL} then element is not added to the data frame.}
}
\value{
A \code{tibble} containing the necessary data to run Comet or other
analyses.
}
\description{
Given a Nimue or Squire simulation object with MCMC fits, this function
produces a data frame of Rt over time and the equivalent Beta values, and if
an Nimue object, vaccine related parameters. It also extends this data frame
for the next 240 days for use in forecasting.
}
\details{
Currently can only take fits to reported deaths and not excess mortality.
}
\examples{
#apply this process to the standard Afghanistan fit
df <- prepare_input_json_df(
   afg_fit,
   ox_interventions = NULL
   )
#can be written to a json as needed
\dontrun{
   writeLines(
      jsonlite::toJSON(df, pretty = TRUE),
      "input_params.json"
      )
}
}
