% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rt_plot_immunity}
\alias{rt_plot_immunity}
\title{Plot the Rt trends accounting for immunity}
\usage{
rt_plot_immunity(out, vaccine = TRUE, R0_plot = FALSE, Rt_plot = FALSE)
}
\arguments{
\item{out}{model output}

\item{vaccine}{Should we account for vaccinations?}

\item{R0_plot}{plot R0}

\item{Rt_plot}{plot Rt}
}
\description{
Plot the Rt trends accounting for immunity
}
