% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booster_model.R
\name{run_booster}
\alias{run_booster}
\title{Run the vaccine model}
\usage{
run_booster(
  country = NULL,
  population = NULL,
  tt_contact_matrix = 0,
  contact_matrix_set = NULL,
  R0 = 3,
  tt_R0 = 0,
  beta_set = NULL,
  time_period = 365,
  replicates = 10,
  seed = stats::runif(1, 0, 1e+08),
  prob_hosp = probs_booster$prob_hosp,
  prob_hosp_multiplier = probs_booster$prob_hosp_multiplier,
  tt_prob_hosp_multiplier = probs_booster$tt_prob_hosp_multiplier,
  prob_severe = probs_booster$prob_severe,
  prob_severe_multiplier = probs_booster$prob_severe_multiplier,
  tt_prob_severe_multiplier = probs_booster$tt_prob_severe_multiplier,
  prob_non_severe_death_treatment = probs_booster$prob_non_severe_death_treatment,
  prob_non_severe_death_no_treatment = probs_booster$prob_non_severe_death_no_treatment,
  prob_severe_death_treatment = probs_booster$prob_severe_death_treatment,
  prob_severe_death_no_treatment = probs_booster$prob_severe_death_no_treatment,
  rel_infectiousness = probs_booster$rel_infectiousness,
  rel_infectiousness_vaccinated = probs_booster$rel_infectiousness_vaccinated,
  dur_E = durs_booster$dur_E,
  tt_dur_E = durs_booster$tt_dur_E,
  dur_IMild = durs_booster$dur_IMild,
  tt_dur_IMild = durs_booster$tt_dur_IMild,
  dur_ICase = durs_booster$dur_ICase,
  tt_dur_ICase = durs_booster$tt_dur_ICase,
  dur_get_ox_survive = durs_booster$dur_get_ox_survive,
  tt_dur_get_ox_survive = durs_booster$tt_dur_get_ox_survive,
  dur_get_ox_die = durs_booster$dur_get_ox_die,
  tt_dur_get_ox_die = durs_booster$tt_dur_get_ox_die,
  dur_not_get_ox_survive = durs_booster$dur_not_get_ox_survive,
  dur_not_get_ox_die = durs_booster$dur_not_get_ox_die,
  dur_get_mv_survive = durs_booster$dur_get_mv_survive,
  tt_dur_get_mv_survive = durs_booster$tt_dur_get_mv_survive,
  dur_get_mv_die = durs_booster$dur_get_mv_die,
  tt_dur_get_mv_die = durs_booster$tt_dur_get_mv_die,
  dur_not_get_mv_survive = durs_booster$dur_not_get_mv_survive,
  dur_not_get_mv_die = durs_booster$dur_not_get_mv_die,
  dur_rec = durs_booster$dur_rec,
  dur_R = vaccine_pars_booster$dur_R,
  tt_dur_R = vaccine_pars_booster$tt_dur_R,
  dur_V = vaccine_pars_booster$dur_V,
  tt_dur_V = vaccine_pars_booster$tt_dur_V,
  vaccine_efficacy_infection = vaccine_pars_booster$vaccine_efficacy_infection,
  tt_vaccine_efficacy_infection = vaccine_pars_booster$tt_vaccine_efficacy_infection,
  vaccine_efficacy_disease = vaccine_pars_booster$vaccine_efficacy_disease,
  tt_vaccine_efficacy_disease = vaccine_pars_booster$tt_vaccine_efficacy_disease,
  primary_doses = vaccine_pars_booster$primary_doses,
  tt_primary_doses = vaccine_pars_booster$tt_primary_doses,
  booster_doses = vaccine_pars_booster$booster_doses,
  tt_booster_doses = vaccine_pars_booster$tt_booster_doses,
  second_dose_delay = vaccine_pars_booster$second_dose_delay,
  vaccine_coverage_mat = vaccine_pars_booster$vaccine_coverage_mat,
  vaccine_booster_initial_coverage =
    vaccine_pars_booster$vaccine_booster_initial_coverage,
  vaccine_booster_follow_up_coverage =
    vaccine_pars_booster$vaccine_booster_follow_up_coverage,
  protection_delay_rate = vaccine_pars_booster$protection_delay_rate,
  protection_delay_shape = vaccine_pars_booster$protection_delay_shape,
  hosp_bed_capacity = NULL,
  ICU_bed_capacity = NULL,
  tt_hosp_beds = 0,
  tt_ICU_beds = 0,
  seeding_cases = 20,
  seeding_age_order = NULL,
  init = NULL,
  use_dde = TRUE,
  ...
)
}
\arguments{
\item{country}{Character for country beign simulated. WIll be used to
generate \code{population} and \code{contact_matrix_set} if
unprovided. Either \code{country} or \code{population} and
\code{contact_matrix_set} must be provided.}

\item{population}{Population vector (for each age group). Default = NULL,
which will cause population to be sourced from \code{country}}

\item{tt_contact_matrix}{Time change points for matrix change. Default = 0}

\item{contact_matrix_set}{Contact matrices used in simulation. Default =
NULL, which will generate this based on the \code{country}.}

\item{R0}{Basic Reproduction Number. Default = 3}

\item{tt_R0}{Change time points for R0. Default = 0}

\item{beta_set}{Alternative parameterisation via beta rather than R0.
Default = NULL, which causes beta to be estimated from R0}

\item{time_period}{Length of simulation. Default = 365}

\item{replicates}{Number of replicates. Default = 10}

\item{seed}{Random seed used for simulations. Deafult = runif(1, 0, 10000)}

\item{prob_hosp}{probability of hospitalisation by age.
Default = c(0.000744192, 0.000634166,0.001171109, 0.002394593, 0.005346437,
0.010289885, 0.016234604, 0.023349169, 0.028944623, 0.038607042,
0.057734879, 0.072422135, 0.101602458, 0.116979814, 0.146099064,
0.176634654 ,0.180000000)}

\item{prob_hosp_multiplier}{Time varying multiplier to probability of developing
severe symptoms. Default = 1, which is no change to provided prob_hosp.}

\item{tt_prob_hosp_multiplier}{Timing of changes to multiplier of probability of
developing severe symptoms. Default = 0}

\item{prob_severe}{Probability of developing severe symptoms by age.
Default = c(0.05022296,    0.05022296,    0.05022296,    0.05022296,    0.05022296,
0.05022296,    0.05022296,    0.053214942, 0.05974426,    0.074602879,
0.103612417, 0.149427991, 0.223777304,    0.306985918,
0.385779555, 0.461217861, 0.709444444)}

\item{prob_severe_multiplier}{Time varying multiplier to probability of
hospitalisation. Default = 1, which is no change to provided prob_hosp.}

\item{tt_prob_severe_multiplier}{Timing of changes to multiplier of probability
of hospitalisation. Default = 0}

\item{prob_non_severe_death_treatment}{Probability of death from non severe
treated infection.
Default = c(0.0125702,    0.0125702,    0.0125702,    0.0125702,
0.0125702,    0.0125702,    0.0125702,    0.013361147,
0.015104687,    0.019164124,    0.027477519,    0.041762108,
0.068531658,    0.105302319,    0.149305732,    0.20349534,    0.5804312)}

\item{prob_non_severe_death_no_treatment}{Probability of death in non severe
hospital inections that aren't treated}

\item{prob_severe_death_treatment}{Probability of death from severe infection
that is treated. Default = rep(0.5, 17)}

\item{prob_severe_death_no_treatment}{Probability of death from severe infection
that is not treated. Default = rep(0.95, 17)}

\item{rel_infectiousness}{Relative infectiousness per age category relative
to maximum infectiousness category. Default = rep(1, 17)}

\item{rel_infectiousness_vaccinated}{Relative infectiousness per age
category  of vaccinated individuals relative to unvaccinated individuals.
Default = rep(1, 17), which is no impact of vaccination on onwards
transmissions}

\item{dur_E}{Mean duration of incubation period (days). Default = 4.6}

\item{tt_dur_E}{Times at which dur_E changes, default = 0.}

\item{dur_IMild}{Mean duration of mild infection (days). Default = 2.1}

\item{tt_dur_IMild}{Times at which dur_IMild changes, default = 0.}

\item{dur_ICase}{Mean duration from symptom onset to hospital admission (days).
Default = 4.5}

\item{tt_dur_ICase}{Times at which dur_ICase changes, default = 0.}

\item{dur_get_ox_survive}{Mean duration of oxygen given survive. Default = 5. Can be
time varying, with timing of changes given by tt_dur_get_ox_survive.}

\item{tt_dur_get_ox_survive}{Timing of changes in duration of  oxygen given survive.}

\item{dur_get_ox_die}{Mean duration of oxygen given death. Default = 5. Can be
time varying, with timing of changes given by tt_dur_get_ox_die.}

\item{tt_dur_get_ox_die}{Timing of changes in duration of  oxygen given death.}

\item{dur_not_get_ox_survive}{Mean duration without oxygen given survive.
Default = 5}

\item{dur_not_get_ox_die}{Mean duration without  oxygen given death.
Default = 5}

\item{dur_get_mv_survive}{Mean duration of ventilation given survive.
Default = 7.3. Can be time varying, with timing of changes given by tt_dur_get_mv_survive.}

\item{tt_dur_get_mv_survive}{Timing of changes in duration of ventilation given survive.}

\item{dur_get_mv_die}{Mean duration of ventilation given death. Default = 6. Can be
time varying, with timing of changes given by tt_dur_get_mv_die.}

\item{tt_dur_get_mv_die}{Timing of changes in duration of ventilation given death.}

\item{dur_not_get_mv_survive}{Mean duration without ventilation given
survive. Default = 7.3}

\item{dur_not_get_mv_die}{Mean duration without ventilation given
death. Default = 1}

\item{dur_rec}{Duration of recovery after coming off ventilation. Default = 2}

\item{dur_R}{Mean duration of naturally acquired immunity (days). Can be
time varying, with timing of changes given by tt_dur_R.}

\item{tt_dur_R}{Timing of changes in duration of natural immunity.}

\item{dur_V}{Mean duration of vaccine-derived immunity (days) for partial protection and full protection. Should be a
numeric vector of length 3, corresponding to the duration of time in each waned compartmenet after recieving a first dose and then for the two second dose compartments.
Alternatively can be a list of values if this changes over time.}

\item{tt_dur_V}{List of change times for dur_V.}

\item{vaccine_efficacy_infection}{Efficacy of vaccine against infection.
This parameter must either be a length 6 numeric (a single efficacy for
each vaccine state (first dose, second dose, waned second dose, booster dose, and two waned second dose compartments))
or numeric vector with 17 columns and 6 rows
(an efficacy for each age group and vaccine state).
An efficacy of 1 will reduce FOI by 100 percent, an efficacy of 0.2 will
reduce FOI by 20 percent etc.
To specify changes in vaccine efficacy over time, vaccine efficacies must
be provided as a list, with each list element being the efficacy at each
time point specified by \code{tt_vaccine_efficacy_infection}. These
efficacies must also be length 6 numeric or 6x17 numeric matrix.}

\item{tt_vaccine_efficacy_infection}{Timing of changes in vaccine efficacy
against infection. Default = 0, which assumes fixed efficacy over time.
Must be the same length as the length of \code{vaccine_efficacy_infection}
when provided as a list. Time changing efficacies can occur in response to
changing vaccines being  given and dosing strategy changes.}

\item{vaccine_efficacy_disease}{Efficacy of partial vaccination against severe
(requiring hospitilisation) disease (by age). This parameter must either be
length 6 numeric (a single efficacy for each vaccine state (first dose, second dose, waned second dose, booster dose, and two waned second dose compartments)) or numeric vector with 17 columns and 6 rows
(an efficacy for each age group and vaccine state). An efficacy of 1 will
reduce the probability of hospitalisation by 100 percent, an efficacy of
0.2 will reduce the probability of hospitalisation by 20 percent etc.
To specify changes in vaccine efficacy over time, vaccine efficacies must
be provided as a list, with each list element being the efficacy at each
time point specified by \code{tt_vaccine_efficacy_disease}. These
efficacies must also be length 6 numeric or 5x17 numeric matrix.}

\item{tt_vaccine_efficacy_disease}{Timing of changes in vaccine efficacy
against disease. Default = 0, which assumes fixed efficacy over time.
Must be the same length as the length of \code{vaccine_efficacy_disease}
when provided as a list. Time changing efficacies can occur in response to
changing vaccines being  given and dosing strategy changes.}

\item{primary_doses}{The maximum number of individuals who can be vaccinated with their first dose per day.}

\item{tt_primary_doses}{Time change points for vaccine capacity (\code{first_doses}).}

\item{booster_doses}{The maximum number of individuals who can be vaccinated with their booster dose per day.}

\item{tt_booster_doses}{Time change points for vaccine capacity (\code{booster_doses}).}

\item{second_dose_delay}{Delay between first dose and second dose in the initial series, this model assume all who get first doses get a second dose, default = 60 (days).}

\item{vaccine_coverage_mat}{Vaccine coverage targets by age (columns) and priority (row)}

\item{vaccine_booster_initial_coverage}{Age group eligibility for initial boosters (i.e. 1st booster dose),
default = NULL means all are eligible. Format: 0 indicates not eligible, 1 indicates eligible.}

\item{vaccine_booster_follow_up_coverage}{Age group eligibility for follow-up boosters (i.e. 2nd, 3rd, ... booster doses),
default = NULL means all are eligible. Format: 0 indicates not eligible, 1 indicates eligible.}

\item{protection_delay_rate}{Rate for the delay in development of vaccine protection, applied via gamma/erlang distribution,
default = 1/14. If NULL no delay is applied.}

\item{protection_delay_shape}{Shape for the delay in development of vaccine protection, applied via gamma/erlang distribution,
default = 1/14. If NULL no delay is applied.}

\item{hosp_bed_capacity}{General bed capacity. Can be single number of vector if capacity time-varies.}

\item{ICU_bed_capacity}{ICU bed capacity. Can be single number of vector if capacity time-varies.}

\item{tt_hosp_beds}{Times at which hospital bed capacity changes (Default = 0 = doesn't change)}

\item{tt_ICU_beds}{Times at which ICU bed capacity changes (Default = 0 = doesn't change)}

\item{seeding_cases}{Initial number of cases seeding the epidemic}

\item{seeding_age_order}{Vector specifying the order in which seeds are allocated to ages.
If NULL, seeds are distributed randomly within working ages. If specified, must be a vector
of length 17 specifying the order seeds are allocated, e.g. 1:17 will allocate first seed
to the youngest age group, then the second youngest and so on. Default = NULL}

\item{init}{Initial conditions for simulation provided. Allows overriding
if initial conditions start with an already infected population etc.
Default = NULL.}

\item{use_dde}{Use the dde solver (default is \code{TRUE})}

\item{...}{Additional arguments for solver}
}
\value{
Simulation output
}
\description{
Run the vaccine model
}
