% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sero_plot}
\alias{sero_plot}
\title{Plot the seroprevalence from the model
Under some assumptions}
\usage{
sero_plot(res, sero_df)
}
\arguments{
\item{res}{model output}

\item{sero_df}{Data to plot against}
}
\description{
Plot the seroprevalence from the model
Under some assumptions
}
