% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{setup_parameters.booster_min_model}
\alias{setup_parameters.booster_min_model}
\title{If no particular method we default to calling the parameter function
attached to the model}
\usage{
\method{setup_parameters}{booster_min_model}(model_obj, parameters)
}
\arguments{
\item{model_obj}{A nimue/squire model object}

\item{parameters}{The parameters to the pass to the parameter function}
}
\description{
If no particular method we default to calling the parameter function
attached to the model
}
