% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squire.page-package.R
\docType{package}
\name{squire.page-package}
\alias{squire.page}
\alias{squire.page-package}
\title{squire.page: Functions used in global-lmic-reports-orderly and other repos}
\description{
Contains miscalanious functions used in the repo global-lmic-reports-orderly and other packages that utilise squire or nimue fits. These functions are kept externally to keep them consistent across mutliple repositories and multiple orderly tasks.
}
\author{
\strong{Maintainer}: Gregory Barnsley \email{g.barnsley@imperial.ac.uk} (\href{https://orcid.org/0000-0001-7972-0737}{ORCID})

}
\keyword{internal}
