% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_optimise_processing.R
\name{trim_rt_optimise}
\alias{trim_rt_optimise}
\title{Trim an Rt Optimised output}
\usage{
trim_rt_optimise(out, p_error)
}
\arguments{
\item{out}{An object of type rt_optimised.}

\item{p_error}{Cutoff value, see description.}
}
\value{
An object of type rt_optimised_trimmed, (model type).
}
\description{
Removes poorly fitted trajectories based upon a predefined cutoff.
For each optimised trajectory this calculates the least squares error from
the observed data, then scales those errors so that the error for estimating
0's is 1 then rejects any trajectories with an error higher than the p_error
value.
}
