% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_fit_internal.R
\name{update_initial_state}
\alias{update_initial_state}
\title{Update initial state with model output}
\usage{
update_initial_state(initial_state, model_output)
}
\description{
Update initial state with model output
}
