% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections_estimate_scenarios.R
\name{update_Rt}
\alias{update_Rt}
\title{Applies new Rt to the data}
\usage{
update_Rt(model_user_args, Rt, model_out)
}
\arguments{
\item{model_user_args}{list of values for each replicates}

\item{Rt}{Values to change Rt to}

\item{model_out}{model object}
}
\description{
Applies new Rt to the data
}
