% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_draws}
\alias{generate_draws}
\title{Generate Draws from pmcmc run}
\usage{
generate_draws(out, pars.list = NULL, draws = 10)
}
\arguments{
\item{out}{Output of \code{squire::pmcmc}}

\item{pars.list}{Output of generate_parameters(), default = NULL calls
generate_parameters.}

\item{draws}{Number of draws from mcmc chain. Default = 10, if NULL then uses
replicate_parameters as the parameters}
}
\description{
Uses furrr, so can be called in parallel.
}
