% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_draws}
\alias{generate_draws}
\title{Generate Draws from a model fit.}
\usage{
generate_draws(out, ...)
}
\arguments{
\item{out}{Output from a fitted MCMC or Rt Optimise}

\item{...}{method specific arguments.}
}
\description{
Appends the simulated values to \code{$output}.
}
