% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_draws.rt_optimised}
\alias{generate_draws.rt_optimised}
\title{Generate Draws from pmcmc run}
\usage{
\method{generate_draws}{rt_optimised}(out, t_end = NULL, project_forwards = FALSE, ...)
}
\arguments{
\item{out}{Output of \code{squire::rt_optimised}}

\item{t_end}{Time to simulate up to}

\item{project_forwards}{Should we use the previous outputs as our start point
and treat this as projection, default = FALSE}

\item{...}{method specific arguments, unused}
}
\description{
Generate Draws from pmcmc run
}
