% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_draws_counterfactual}
\alias{generate_draws_counterfactual}
\title{Generate Draws from pmcmc run with a counter factual based on vaccines}
\usage{
generate_draws_counterfactual(
  out,
  pars.list,
  assignedVaccine,
  vaccineStart = NULL,
  draws = 10
)
}
\arguments{
\item{out}{Output of \code{squire::pmcmc}}

\item{pars.list}{Output of generate_parameters(), default = NULL calls
generate_parameters.}

\item{assignedVaccine}{The number of doses given to the country at this date.}

\item{vaccineStart}{The date the vaccination is to start. Should be a character
in the format "YYYY-MM-DD". Default = NULL}

\item{draws}{Number of draws from mcmc chain. Default = 10, if NULL then uses
replicate parameters.}
}
\description{
Uses furrr, so can be called in parallel.
}
