% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop_conv.R
\name{get_dates.particle_fit}
\alias{get_dates.particle_fit}
\title{S3 Method to get all the start dates of time-periods in the data of a model}
\usage{
\method{get_dates}{particle_fit}(model_out)
}
\description{
S3 Method to get all the start dates of time-periods in the data of a model
}
