% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fits_from_orderly.R
\name{get_fits}
\alias{get_fits}
\title{Get the relevant fits from an global-lmic-reports-orderly.
Uses DBI to search orderlys database for the relevant files.
Save the found model objects in ~/model_fits/... under the
relevant iso3 code.}
\usage{
get_fits(repo, iso3cs, date)
}
\arguments{
\item{repo}{the file path of the global-lmic-reports-orderly repo.}

\item{iso3cs}{which countries should be collected?}

\item{date}{Collect the reports from which date, should be a string}
}
\description{
Get the relevant fits from an global-lmic-reports-orderly.
Uses DBI to search orderlys database for the relevant files.
Save the found model objects in ~/model_fits/... under the
relevant iso3 code.
}
