% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fits_from_orderly.R
\name{get_fits}
\alias{get_fits}
\title{Get the relevant fits from an global-lmic-reports-orderly}
\usage{
get_fits(repo, date, iso3cs = NULL, excess = FALSE)
}
\arguments{
\item{repo}{The file path of the global-lmic-reports-orderly repo.}

\item{date}{Collect the reports from which date, should be a string.}

\item{iso3cs}{Which countries should be collected? Default = NULL, collects all available.}

\item{excess}{Collect excess fits? Default = FALSE.}
}
\value{
A named (iso3 codes) list of fit objects.
}
\description{
Uses DBI to search orderlys database for the relevant files. Also requires DBI
search the orderly database. Returns a list of model fits, warning might be
large, since this is only run on my PC with large amounts of RAM, this is fine.
Also \code{iso3cs} can be used to reduce the list size.
}
\examples{
#get fits from my orderly repo
\dontrun{
get_fits(
   "C:\\\\Users\\\\gbarnsle\\\\Documents\\\\Covid Vaccine Impact\\\\global-lmic-reports-orderly",
   "2021-11-21",
   iso3cs = NULL,
   excess = TRUE
)
}
}
