% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_load.R
\name{loadCounterfactualData}
\alias{loadCounterfactualData}
\title{A function to load and summarise the deaths averted data}
\usage{
loadCounterfactualData(
  counterfactuals,
  group_by,
  quantileSamples = 2000,
  exclude_iso3cs = NULL
)
}
\arguments{
\item{counterfactuals}{Which counterfactuals to load. A vector of strings.
Since these are large files, it is recommended to load only one counterfactual
at a time.}

\item{group_by}{A vector of strings of the variables to group the summary by,
e.g. Iso3c code, age_group or date.}

\item{quantileSamples}{Only used if dataset it not grouped by country. The
number of combinations to randomly draw. Defaults to 2000. If NULL creates
a wider interval based on ordering samples.}

\item{exclude_iso3cs}{A vector of iso3cs to exclude from the summaries,
default = NULL.}
}
\description{
A function to load and summarise the deaths averted data
}
