% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimue_utils.R
\name{nimue_format}
\alias{nimue_format}
\title{Nimue format that returns just compartments/summaries requested}
\usage{
nimue_format(
  out,
  var_select = NULL,
  reduce_age = TRUE,
  combine_compartments = TRUE,
  date_0 = NULL,
  mean_over_80_age = 82.5,
  case_to_infection_ratio = 0.195
)
}
\arguments{
\item{out}{nimue_simulation object}

\item{var_select}{Vector of compartment names, e.g. \code{c("S", "R")}. In
addition a number of summary compartment can be requested. These include:
\itemize{
\item{"deaths"}{ Daily Deaths }
\item{"infections"}{ Daily Infections }
\item{"hospital_occupancy"}{ Occupied Hospital Beds }
\item{"ICU_occupancy"}{ Occupied ICU Beds }
\item{"hospital_demand"}{ Required Hospital Beds }
\item{"ICU_demand"}{ Required ICU Beds }
\item{"hospital_incidence"}{ Incidence of hospitilisation }
\item{"ICU_incidence"}{ Incidence of ICU admissions }
\item{"hospitilisations"}{ Incidence of hospital+ICU admissions }
\item{""vaccines", "unvaccinated", "vaccinated", "priorvaccinated"}{ Vaccine outputs }
}}

\item{reduce_age}{Collapse age-dimension, calculating the total in the
compartment.}

\item{combine_compartments}{Collapse compartments of same type together
(e.g. E1 and E2 -> E)}

\item{date_0}{Date of time 0, if specified a date column will be added}

\item{mean_over_80_age}{Mean age of over 80 population for calculating long
covid incidence. Default = 82.5}

\item{case_to_infection_ratio}{Ratio of infections to cases for long covid
calculation. Default = 0.195 from UK ONS analyses}
}
\value{
Formatted long data.frame
}
\description{
Nimue format that returns just compartments/summaries requested
}
