% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan_fit}
\alias{scan_fit}
\title{Fit a nimue model to the given data using a scan fitting method}
\usage{
scan_fit(
  res,
  data = NULL,
  width = 2,
  n_span = 8,
  width_end = NULL,
  repeats = 5
)
}
\arguments{
\item{res}{A nimue simulation object}

\item{data}{The data to fit to, for standard fits must have date and death
columns; for lmic fits must have date, deaths and cases; and for excess
mortality it must have deaths, week_start and week_end. Default = NULL, which uses the data in the fit}

\item{width}{The algorithm will check values up to +- width of the current
value}

\item{n_span}{How many values between the width to check}

\item{width_end}{The final width to check, so that the algorithm to move to
a finer resolution as the proceeds, default NULL means the same as width}

\item{repeats}{How many times to iterate through the process}
}
\value{
Model object tuned to the data
}
\description{
Taken from OJs code. Iteratively adjusts the Rt_rw parameters based on
maximising likelihood of a short period of time in the data.
}
\details{
Only works with excess fits for now.

Uses furrr for simulations later so can be run in parallel with
future::plan(future::multisession(), .cleanup = TRUE)
}
